modeling_inferServer/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.5.0
# Editable install with no version control (anomalib==0.6.0)
-e /home/i4624/.local/lib/python3.10/site-packages
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argumentparser==1.2.1
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.4
blis==0.7.10
boto3==1.28.15
botocore==1.31.15
Brlapi==0.8.3
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.4
colorful==0.5.5
comm==0.1.3
command-not-found==0.3
confection==0.1.0
constantly==15.1.0
construct==2.10.68
contourpy==1.1.0
coqpit==0.0.17
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dask==2023.7.1
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distributed==2023.7.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.4.0
docker-pycreds==0.4.0
docstring-parser==0.15
duplicity==0.8.21
einops==0.6.1
email-validator==2.0.0.post2
emoji==2.6.0
empath==0.89
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl#sha256=9012584a8fcf5bc61bedb13a4b511fb81ffff9f770958ee06366671be10632e9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
fasteners==0.14.1
fastjsonschema==2.18.0
filelock==3.12.2
flashtext==2.7
flatbuffers==23.5.26
fonttools==4.41.0
fqdn==1.5.1
FrEIA==0.2
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
future==0.18.2
gast==0.4.0
gdown==4.7.1
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
gpustat==1.1
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httplib2==0.20.2
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==3.3
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
incremental==22.10.0
inexactsearch==1.0.2
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
json5==0.9.14
jsonargparse==4.22.1
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.13.1
keyring==23.5.0
kiwisolver==1.4.4
kornia==0.6.9
langcodes==3.3.0
langdetect==1.0.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
lit==16.0.6
locket==1.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
modcall==0.1.0
modin==0.23.0
modin-spreadsheet==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpi4py-mpich==3.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
namespace==0.1.4
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
nn==0.1.1
notebook==7.0.0
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
openai==0.27.8
opencensus==0.11.2
opencensus-context==0.1.3
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
openvino==2023.0.1
opt-einsum==3.3.0
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
partd==1.4.0
path==16.7.1
pathtools==0.1.2
pathy==0.10.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
plac==1.1.3
platformdirs==3.9.1
pluggy==1.2.0
plumbum==1.8.2
ply==3.11
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycld2==0.41
pycparser==2.21
pycryptodomex==3.18.0
pycups==2.0.1
pydantic==1.10.11
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pyLDAvis==3.4.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pysmi==0.3.4
pysnmp==4.4.12
PySocks==1.7.1
pytest==7.4.0
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==1.9.5
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
ray==2.6.1
referencing==0.30.0
regex==2023.6.3
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rpyc==4.1.5
rsa==4.9
s3transfer==0.6.1
scattertext==0.1.19
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==1.28.1
setproctitle==1.3.2
shapely==2.0.1
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundex==1.1.3
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.6.0
spacy-cld==0.1.0
spacy-langdetect==0.1.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacymoji==3.1.0
spellchecker==0.4
srsly==2.4.7
ssh-import-id==5.11
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
stopwords==1.0.0
sympy==1.12
systemd-python==234
tblib==2.0.0
tensor==0.3.6
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.6.12
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchinfo==1.8.0
torchmetrics==0.10.3
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
trainer==0.0.28
traitlets==5.9.0
triton==2.0.0
Twisted==22.10.0
typer==0.9.0
typeshed-client==2.3.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
unidist==0.4.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
usb-creator==0.3.7
uvicorn==0.23.1
uvloop==0.17.0
virtualenv==20.21.0
wadllib==1.3.6
wandb==0.15.5
wasabi==0.10.1
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wordcloud==1.9.2
wrapt==1.15.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zict==3.0.0
zipp==1.0.0
zope.interface==6.0
